Prepare data.indexPatterns for move to NP (#48083)
Replace ui/saved_objects by core.savedObjects
