+++ src/legacy/core_plugins/data/public/index_patterns/index_patterns/index_pattern.tsx
@@ -27,10 +27,10 @@ import chrome from 'ui/chrome';
 import { fieldFormats } from 'ui/registry/field_formats';
 // @ts-ignore
 import { expandShorthand } from 'ui/utils/mapping_setup';
-import { findObjectByTitle } from 'ui/saved_objects';
+
 import { NotificationsSetup, SavedObjectsClientContract } from 'src/core/public';
 import { SavedObjectNotFound, DuplicateField } from '../../../../../../plugins/kibana_utils/public';
-
+import { findIndexPatternByTitle } from '../utils';
 import { IndexPatternMissingIndices } from '../errors';
 import { Field, FieldList, FieldType, FieldListInterface } from '../fields';
 import { createFieldsFetcher } from './_fields_fetcher';
@@ -469,9 +469,8 @@ export class IndexPattern implements StaticIndexPattern {
       return response.id;
     };
 
-    const potentialDuplicateByTitle = await findObjectByTitle(
+    const potentialDuplicateByTitle = await findIndexPatternByTitle(
       this.savedObjectsClient,
-      type,
       this.title
     );
     // If there is potentially duplicate title, just create it
