+++ src/legacy/core_plugins/data/public/index_patterns/index_patterns/index_patterns.test.ts
@@ -36,12 +36,6 @@ jest.mock('ui/registry/field_formats', () => ({
   },
 }));
 
-jest.mock('ui/notify', () => ({
-  toastNotifications: {
-    addDanger: jest.fn(),
-  },
-}));
-
 jest.mock('./index_pattern', () => {
   class IndexPattern {
     init = async () => {
