+++ src/legacy/core_plugins/data/public/index_patterns/utils.ts
@@ -17,11 +17,13 @@
  * under the License.
  */
 
-import { get } from 'lodash';
+import { find, get } from 'lodash';
 
 import { Field, FieldType } from './fields';
 import { StaticIndexPattern } from './index_patterns';
 import { getFilterableKbnTypeNames } from '../../../../../plugins/data/public';
+
+import { SavedObjectsClientContract, SimpleSavedObject } from '../../../../../core/public';
 
 export const ILLEGAL_CHARACTERS = 'ILLEGAL_CHARACTERS';
 export const CONTAINS_SPACES = 'CONTAINS_SPACES';
@@ -42,6 +44,48 @@ function findIllegalCharacters(indexPattern: string): string[] {
   );
 
   return illegalCharacters;
+}
+
+/**
+ * Returns an object matching a given title
+ *
+ * @param client {SavedObjectsClientContract}
+ * @param title {string}
+ * @returns {Promise<SimpleSavedObject|undefined>}
+ */
+export async function findIndexPatternByTitle(
+  client: SavedObjectsClientContract,
+  title: string
+): Promise<SimpleSavedObject<any> | void> {
+  if (!title) {
+    return Promise.resolve();
+  }
+
+  const { savedObjects } = await client.find({
+    type: 'index-pattern',
+    perPage: 10,
+    search: `"${title}"`,
+    searchFields: ['title'],
+    fields: ['title'],
+  });
+
+  return find(
+    savedObjects,
+    (obj: SimpleSavedObject<any>) => obj.get('title').toLowerCase() === title.toLowerCase()
+  );
+}
+
+export async function getIndexPatternTitle(
+  client: SavedObjectsClientContract,
+  indexPatternId: string
+): Promise<SimpleSavedObject<any>> {
+  const savedObject = (await client.get('index-pattern', indexPatternId)) as SimpleSavedObject<any>;
+
+  if (savedObject.error) {
+    throw new Error(`Unable to get index-pattern title: ${savedObject.error.message}`);
+  }
+
+  return savedObject.attributes.title;
 }
 
 function indexPatternContainsSpaces(indexPattern: string): boolean {
