+++ src/legacy/core_plugins/vis_type_vega/public/vega_visualization.js
@@ -21,9 +21,10 @@ import { i18n } from '@kbn/i18n';
 import { toastNotifications } from 'ui/notify';
 import { VegaView } from './vega_view/vega_view';
 import { VegaMapView } from './vega_view/vega_map_view';
-import { findObjectByTitle } from 'ui/saved_objects';
 import { timefilter } from 'ui/timefilter';
 import { start as data } from '../../../core_plugins/data/public/legacy';
+
+import { findIndexPatternByTitle } from '../../data/public/index_patterns';
 
 export const createVegaVisualization = ({ serviceSettings }) => class VegaVisualization {
   constructor(el, vis) {
@@ -40,7 +41,7 @@ export const createVegaVisualization = ({ serviceSettings }) => class VegaVisual
   async findIndex(index) {
     let idxObj;
     if (index) {
-      idxObj = await findObjectByTitle(this.savedObjectsClient, 'index-pattern', index);
+      idxObj = await findIndexPatternByTitle(this.savedObjectsClient, index);
       if (!idxObj) {
         throw new Error(i18n.translate('visTypeVega.visualization.indexNotFoundErrorMessage', {
           defaultMessage: 'Index {index} not found',
